Rename to launches to keep consistency
